+++ src/main/java/com/example/rentalSystem/global/response/type/SuccessType.java
@@ -1,0 +1,17 @@
+package com.example.rentalSystem.global.response.type;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@RequiredArgsConstructor
+public enum SuccessType {
+
+    SUCCESS(200, "요청에 성공하였습니다."),
+    CREATED(201, "성공적으로 생성하였습니다."),
+    NO_CONTENT(204, "요청에 성공하였습니다.");
+
+    private final int httpStatusCode;
+    private final String message;
+}
